Inspect: Add error if no results (#19476)
When no results match or another error occurs, add an error to the stream. Keep the "inspect-input.txt" as the only thing in the zip for reference.

Example:

```
λ mc support inspect --airgap myminio/testbucket/fjghfjh/**
mc: Using public key from C:\Users\klaus\mc\support_public.pem
File data successfully downloaded as inspect-data.enc

λ inspect inspect-data.enc
Using private key from support_private.pem
output written to inspect-data.zip
2024/04/11 14:10:51 next stream: GetRawData: No files matched the given pattern

λ unzip -l inspect-data.zip
Archive:  inspect-data.zip
  Length      Date    Time    Name
---------  ---------- -----   ----
      222  2024-04-11 14:10   inspect-input.txt
---------                     -------
      222                     1 file

λ
```

Modifies inspect to read until end of stream to report the error.

Bonus: Add legacy commandline params

+++ docs/debugging/inspect/main.go
@@ -24,6 +24,7 @@ import (
 	"crypto/x509"
 	"encoding/json"
 	"encoding/pem"
+	"errors"
 	"flag"
 	"fmt"
 	"io"
@@ -117,18 +118,23 @@ func main() {
 	fatalErr(err)
 
 	// Decrypt the inspect data
+	msg := fmt.Sprintf("output written to %s", outputFileName)
+
 	switch {
 	case *keyHex != "":
-		err = extractInspectV1(*keyHex, input, output)
+		err = extractInspectV1(*keyHex, input, output, msg)
 	case len(privateKey) != 0:
-		err = extractInspectV2(privateKey, input, output)
+		err = extractInspectV2(privateKey, input, output, msg)
 	}
 	output.Close()
 	if err != nil {
-		os.Remove(outputFileName)
+
+		var keep keepFileErr
+		if !errors.As(err, &keep) {
+			os.Remove(outputFileName)
+		}
 		fatalErr(err)
 	}
-	fmt.Println("output written to", outputFileName)
 
 	// Export xl.meta to stdout
 	if *export {
